Fixed losing nameservers when doing a DVD installation (bnc#668034).
r60729 which intended to fix bnc#561349 claims "write resolv.conf
when change from dhcp->static" but in fact writes it always, even if
there was no config at all, overwriting the DHCP nameservers with an
empty list.

Deleted that bit, bringing the patch closer to the original one of
which it was a "backport", r41079 bnc#327074.

svn path=/branches/SuSE-SLE-10-SP4-Branch/network/; revision=63440