AutoYaST: write /etc/hosts even if keep_install_network=true (bnc#796580)
The regression was introduced in SP2 in yast2-network-2.17.166
(2012-01-13) in a fix for bnc#712864.
https://github.com/yast/yast-network/commit/4f67f62d9f16f774fac6870d932a53d1d29d10dc
There, a Lan::Read(`cache) call is added with the intent to merge in the
config provided by linuxrc; it also calls Host::Read, which should
always cache, but wrongly discards the Host::Import'ed data. So a later
Host::Write call thinks there is nothing new to write.
(y2log: "Host.ycp:153 Hosts not modified")

This is an incomplete fix for early testing. It should work
but leaves the code messy: `modified`, `initialized` and
`hosts_init` should all be merged.
(cherry picked from commit 2ddd18503743be8cf3155fbb840406d9c35d66d0)