Updated changelog.
NetHwDetection::Start() used to set up all links which was undesirable
outside the installation (bnc#782283) so we thought we made it
installation specific in commit 3b7a739 when the side effect was moved
into new function SetAllLinksUp and  called from Lan::Read()

But we did not notice that this side effect was needed by a different
caller, inst_netprobe.ycp (bnc#792985).

Now we call SetAllLinksUp in inst_netprobe where it is needed. In
Lan::Read it does not help so it is removed.