Password setup enabled.
This commit adds possibility to specify if password authentization
is required or not.

Before allowing password setup, the Xvnc service was run in
default authentication mode (which currently is "vncauth").
Non-trivial configuration in another YaST module(s) or cmdline
was required to enable vnc connection (or disable password
authentization) in such case.