enclose client into namespace otherwise global namespace is messed
Problem is that even if clients acts like script, it is evaluated in
global namespace and process is still alive, result is that it mess
global namespace. In this particular case it breaks installation because
it adds Yast::FileUtils to global namespace as ::FileUtils and when we
require original FileUtils from ruby, then it collide and raise
exception.