Force copying netconfig from instsys to target. bnc#861078
Network configuration is more powerfull in first stage now. So, if user
configures it, keep it for him.

Moreover, some changes (e.g. in linuxrc -> netdevice) caused that creating
default configuration for first device doesn't work properly anymore.

And last but not least approach to net devices in installation has
changed. Installer should be able to use any of available devices, not
only the first one.