Implemented per interface setup of DHCLIENT_SET_HOSTNAME
there are at least two approaches to achieve this
1) configure as small of devices as possible with respect to
global setup
2) configure all DHCP devices but one to "no" and one to "yes"
independently to the global setup

The second approach seems to be more robust (e.g. not so sensitive
to changes to global option out of yast). Moreover system do not
rely on the global one heavily and uses it really as a default only.