Removed useless code from InstallInfConvertor
Linuxrc does not export some variables anymore handling the network
configuration directly and because of that some code is useless. The
commit responsable of this changes is mainly this one
https://github.com/openSUSE/linuxrc/commits/c07386aedd4458c75753aa534242d49c3974c6cb.

This is the up to date documentation about the linuxrc yast interface
(install.inf file):

https://github.com/openSUSE/linuxrc/blob/master/linuxrc_yast_interface.txt

where you can see that for example Netconfig, Netdevice... are not there.