Fix hash vs keyword arguments in RSpec expectations (bsc#1204871)
New rspec-mocks distinguish between hash and keyword arguments passed to `with`.

The tested code does use hashes but we've used keywords in tests because
it's shorter to write. Fixed.

rspec-mocks 3.11.2 / 2022-10-25:

> Bug Fixes: Support keyword argument semantics when constraining argument
> expectations using with on Ruby 3.0+ with instance_double
> rspec/rspec-mocks#1473

Why this took longer to fix? Misleading RSpec messages:

Even with current rspec-mocks 3.12.0 the error message can be confusing,
differing only in the object_id of the Path:

       Yast::SCR received :Write with unexpected arguments
         expected: (#<Yast::Path:0x00007f92c30a6f88 @components=["modules", "options", "virtio_net"]>, {"debug"=>"16"})
              got: (#<Yast::Path:0x00007f92c30a4d00 @components=["modules", "options", "virtio_net"]>, {"debug"=>"16"})

That's because https://github.com/rspec/rspec-mocks/commit/e931e818b577172b89fb4583fc336fbcd25df36b
only adds the distinguishing "(keyword arguments)" or "(options hash)"
when the _inspect representations_ of the arguments are the same.

To avoid this bug we could override #inspect to omit the #object_id
whenever we override #==.